requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.24.1 